+++ mhcnames/__init__.py
@@ -24,7 +24,7 @@ from .named_allele import NamedAllele
 from .mutation import Mutation
 from .mutant_allele import MutantAllele
 from .alpha_beta_pair import AlphaBetaPair
-from .gene_class import GeneClass
+from .mhc_class import MhcClass
 from .serotype import Serotype
 from .dataframe import dataframe_from_list
 
@@ -36,7 +36,7 @@ __all__ = [
     "AlleleParseError",
     "parse",
     "Gene",
-    "GeneClass",
+    "MhcClass",
     "AlleleGroup",
     "FourDigitAllele",
     "SixDigitAllele",
