+++ mhcnames/serotype.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018. Mount Sinai School of Medicine
+# Copyright (c) 2018-2019. Mount Sinai School of Medicine
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -52,5 +52,3 @@ class Serotype(Gene):
             allele.normalized_string() for allele in self.alleles
         ])
         return d
-
-
